Did some changes to prevent mysterious compiler error message
The error was:
Error:(34, 25) type mismatch;
 found   : JMAP[K,V]
 required: ?C[?E]
Note that implicit conversions are not applicable because they are ambiguous:
 both method ArrowAssoc in object Predef of type [A](self: A)ArrowAssoc[A]
 and method Ensuring in object Predef of type [A](self: A)Ensuring[A]
 are possible conversion functions from JMAP[K,V] to ?C[?E]
  var securityContext = new no.uio.musit.security.SecurityContext
                        ^
(And we haven't created any JMAP, as far as I know)
