+++ build.sbt
@@ -55,28 +55,35 @@ val noPublish = Seq(
 lazy val root = (
   project.in(file("."))
   settings(noPublish)
-  aggregate(common, security, service_core, service_example)
+  aggregate(common_test, common, security, service_core, service_example)
 )
 
 // Base projects used as dependencies
 lazy val common = (
   BaseProject("common")
   settings(noPublish)
-  settings(libraryDependencies ++= playWithPersistenceDependencies)
+  settings(libraryDependencies ++= testablePlayWithPersistenceDependencies)
   settings(scoverageSettings: _*)
-)
+) dependsOn(common_test % "test")
+
+lazy val common_test = (
+  BaseProject("common_test")
+    settings(noPublish)
+    settings(libraryDependencies ++= playWithPersistenceDependencies ++ Seq[ModuleID](scalatestSpec, playframework.specs2Spec))
+    settings(scoverageSettings: _*)
+  )
 
 lazy val security = (
   BaseProject("security")
   settings(noPublish)
-  settings(libraryDependencies ++= playDependencies)
+  settings(libraryDependencies ++= testablePlayDependencies)
   settings(scoverageSettings: _*)
-)  dependsOn(common)
+)  dependsOn(common, common_test % "test")
 
 // Microservices with publish support
 lazy val service_example = (
   PlayProject("service_example")
-  settings(libraryDependencies ++= playWithPersistenceDependencies)
+  settings(libraryDependencies ++= testablePlayWithPersistenceDependencies)
   settings(routesGenerator := InjectedRoutesGenerator)
   settings(scoverageSettings: _*)
   settings(baseDockerSettings ++ Seq(
@@ -86,7 +93,7 @@ lazy val service_example = (
 
 lazy val service_core = (
   PlayProject("service_core")
-    settings(libraryDependencies ++= playWithPersistenceDependencies)
+    settings(libraryDependencies ++= testablePlayWithPersistenceDependencies)
     settings(routesGenerator := InjectedRoutesGenerator)
     settings(scoverageSettings: _*)
     settings(baseDockerSettings ++ Seq(
